token-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
cryptography>=43.0.0
python-multipart>=0.0.9
pydantic>=2.9.0
python-dotenv>=1.0.0
